Refactor: (Feedback) Line 에서는 전달받은 수 만큼 points 생성하도록 수정, RandomPointGenerator 클래스 생성
